Copy assume-role for now
resources aren't available in the generated job's workspace.
We need to either stuff it in with EnvInject, or put it in a git repo
we check out into the workspace all the time.  After that we can
better factor this code and make ourselves happier.

Longer plan is to put it in configuration.git:util/jenkins/
and source it from scripts, but that will be a seperate PR.