Fix overzealous parameter-based throttling
The intention was to prevent the same exact user in the same environment
to have retirement driven concurrently by multiple different retirement
driver jobs in order to prevent race conditions.  This plugin
functionality is clearly broken according to tests, since subsequent
consecutive jobs with DIFFERENT parameters were pending with this
message:

"(pending—A build with matching parameters is already running)"

Replacing the comma separated list with an empty string implies the
matching criteria is all parameters.  This isn't perfect, but it's
better than nothing.