update cron setting for upgrade jobs
Switched cronValue from:
  `@weekly`
to:
  `* H(4-11) * * H(0-4)`

This will choose a random time between midnight-7:59am ET,
Sunday-Thursday, which will have this run mostly off hours (depending on
time zone), and new PRs should be available in the morning during the
week with sufficient time to merge and deploy.

Notes:
- The H symbol can be thought of as a random value over a range, but
it actually is a hash of the job name, not a random function, so that
the value remains stable for any given project.
- Times are in UTC.

For help text for Jenkins cron settings, see:
https://github.com/jenkinsci/jenkins/blob/master/core/src/main/resources/hudson/triggers/TimerTrigger/help-spec.jelly