use cronOffHoursBusinessDaily for edx-platform
Introduced a new variable for off-hours daily upgrade jobs to run
during the week named cronOffHoursBusinessDaily. This is only used for
edx-platform, which would no longer run on Friday or Saturday.

Also renamed the weekly variable to cronOffHoursBusinessWeekday for
clarity.

Also, switched initial `*` to `H` so it no longer runs every minute.