Use pip 20.0.2 when updating dependencies (#953)
The pip-tools 5.0.0 only supports pip>=20, and the previous release won't work with the upcoming pip 20.1. Explicitly install the current pip release to make this upgrade smoother and start gaining confidence that we can safely upgrade to this pip version more broadly. This pip release has been the current one for about 3 months, so it seems unlikely to have any critical bugs.