Fix reference name for Python 3.5 (#1012)
When I updated these from 2.7, I had trouble checking the name of the Python 3.5 installation because the instructions in README-Hacking were incorrect. I presumed it matched the pattern we used for 2.7 (which was recommended by the ShiningPanda Jenkins plugin), but apparently not. I found the correct place in the configuration settings, updated the README, and fixed the broken references. This should get the backup job working again (the OEP-2 checker is currently disabled for other reasons).