Let edx-platform pipeline jobs omit onlyTriggerPhrase (#1031)
The a11y and JS jobs allow `onlyTriggerPhrase` to be omitted (in which case it takes the default `false` value), but this one was requiring it. Change it to match the other two so we don't get tripped up again trying to change this in all the files the same way.