Upgrade repo health repos to Python 3.6 (#1065)
The pytest-repo-health and edx-repo-health repositories were configured to support Python 3.5, but we actually use 3.6 in the Jenkins job. Furthermore, the exact version of 3.5 matters because recent versions of aiohttp dropped support for 3.5.2 and below, and our Jenkins workers still use Ubuntu's 3.5.1. Switching this to 3.6, I'll be updating the Travis configs and minimum supported versions to match.