Remove org from upgrader job names
This reverts one of the changes made in PR #1203/commit 17e2757b50.

Including the org name made it harder to read the listing, and it's
unlikely that we'll have repos of the same name in different orgs that
we also run the upgrader on.

Just in case, add a validator so that we don't end up overwriting a job
if this does happen.