Stop using shiningpanda for user retirement jobs
Shiningpanda was creating an impossible situation for python 3.5 because
it seems to be hard-coded to upgrade pip, but the latest pip (21) has
dropped python 3.5 support.  Also, shiningpanda is no longer supported
upstream, supposedly, so the rest of the edx org has moved away from it.

Additionally, I have changed the python version for these jobs from 3.5
to 3.8.