feat: Turn off Who-Tests-What check until we're ready to work on it again
This check has been running on a daily basis, but the deployment pipeline
has been waiting for it to finish (in check_pr_tests_status.py) meaning
that it sometimes blocks deployment in the mornings as the check is still
in the "pending" state.

We should either delete this check or exclude it from the status checks,
but for now I'm just disabling the daily run on master so that we can defer
that decision.

(I would have used `disabled()` but I don't know if it apples to
`pipelineJob`.)