fix: change emr-cost-reporter virtualenv setup (#1395)
for some reason, the setup script for the emr-cost-reporter has been
failing to create a directory. I'm not sure why, so I am attempting to
change where it is built (WORKSPACE on the mount instead of /tmp on the
root volume) and by giving each virtualenv a unique name by appending
the build id.