fix: provision with frontend-app-learning by default
frontend-app-learning hosts the default course home
and courseware experiences. So, sandboxes are "broken"
(that is, courseware won't load) unless users remember
to check the box to enable frontend-app-learning.

In this fix, we just provision with the MFE by default.
If a dev really doesn't want the MFE, they can uncheck it.

TNL-8535