chore: refactor dbt to use MODEL_SELECTOR instead of MODEL_TAG. (#1477)
for DENG-1070, we need more general selector than just a tag name,
so move the "tag:" prefix into the definition of MODEL_SELECTOR
values.

for DENG-1062, we also need to be able to run 'dbt snapshot' instead
of 'dbt run'.  So define DBT_COMMAND to allow this to be overridden.