chore: load only required seeds using SEED_SELECTOR
In the past, all seeds were loaded on every run of a DBT workflow.
But since all runs use a MODEL_SELECTOR to run a subset of models,
only a subset of seeds are actually needed.  Excluding these helps
streamline the Amplitude-extraction workflow, which writes to
different ("amplitude_"-prefixed) schema names, and some seeds get
loaded into a schema that doesn't exist.  This will be accomplished
by using a non-default value of SEED_SELECTOR, while all other jobs
would continue to use '*' as the selector.

In support of DENG-1070.