feat: Remove Jenkins job that uploads learner data (for MicroBachelors Coaching) to ITK
[MICROBA-1583]

InsideTrack (ITK) has revised their pricing on 1:1 coaching and it has become prohibitively expensive. edX and ITK are parting ways for now. We need to stop uploading learner data to ITK as our contract expires on 01-14-2022. This PR removes the job (and supporting pieces) that uploads learner data to their FTP server.